refactor: separate out presentation preview
